pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "satori"
version = "0.1.0"
description = "Vendor-agnostic CLI tool for evaluating LLM responses using LLM-as-a-Judge approach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.61.0",
    "google-genai>=1.29.0",
    "huggingface-hub>=0.34.3",
    "openai>=1.99.1",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "rich>=14.1.0",
    "ruff>=0.12.7",
    "typer>=0.16.0",
]
authors = [{ name = "Beowolx" }]
keywords = ["llm", "evaluation", "benchmark", "ai", "judge", "cli"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
satori = "src.cli:main"

[project.urls]
Homepage = "https://github.com/beowolx/satori"
Repository = "https://github.com/beowolx/satori"
Issues = "https://github.com/beowolx/satori/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/README.md", "/LICENSE"]

[tool.ruff]
line-length = 80
indent-width = 2
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false
